backend/requirements.txt
Flask
python-dotenv
appwrite
flask-cors
